+++ src/components/forms/index.ts
@@ -1,6 +1,1 @@
-export * from './Input';
-export * from './TextArea';
-export * from './Checkbox';
-export * from './context';
-export * from './Select';
 export * from './radio-group';