+++ src/components/index.ts
@@ -10,7 +10,6 @@ export * from './info-panel/index';
 export * from './layout-manager/index';
 export * from './modal/index';
 export * from './root-layout/index';
-export * from './spinner/index';
 export * from './split-pane/index';
 export * from './table/index';
 export * from './tabs/index';